+++ cognite/powerops/config.py
@@ -604,6 +604,22 @@ class RKOMBidProcessConfig(Configuration):
         return bootstrap_resources
 
 
+class PlantTimeSeriesMapping(BaseModel):
+    plant_name: str
+    water_value: Optional[int]
+    inlet_reservoir_level: Optional[int]
+    outlet_reservoir_level: Optional[int]
+    p_min: Optional[int]
+    p_max: Optional[int]
+    feeding_fee: Optional[int]
+    head_direct: Optional[int]
+
+
+class GeneratorTimeSeriesMapping(BaseModel):
+    generator_name: str
+    start_stop_cost: Optional[int]
+
+
 class BootstrapConfig(BaseModel):
     constants: CommonConstants
     cdf: CDFConfig
@@ -618,6 +634,8 @@ class BootstrapConfig(BaseModel):
     rkom_bid_combination: Optional[list[RKOMBidCombinationConfig]] = None
     rkom_bid_process: list[RKOMBidProcessConfig]
     rkom_market: Optional[RkomMarketConfig] = None
+    plant_time_series_mappings: list[PlantTimeSeriesMapping] = None
+    generator_time_series_mappings: list[GeneratorTimeSeriesMapping] = None
 
     @validator("price_scenario_by_id")
     def no_duplicated_scenarios(cls, value: dict[str, PriceScenario]):