+++ cognite/powerops/data_classes/asset_lists.py
@@ -7,7 +7,7 @@ from cognite.powerops.config import (
     Watercourse,
 )
 from cognite.powerops.data_classes.core import AssetResourceList
-from cognite.powerops.data_classes.generators import Generator
+from cognite.powerops.data_classes.generator import Generator
 from cognite.powerops.data_classes.plant import Plant
 from cognite.powerops.data_classes.price_area import PriceArea
 from cognite.powerops.data_classes.reservoirs import Reservoir